:truck: Move scripts paths
