+++ bin/init.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require('../src/init')();
+require('@umm/scripts').scripts.init();
