+++ bin/module/uninstall.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require('@umm/scripts').scripts.module.uninstall();
