+++ bin/project/deploy.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require('@umm/scripts').scripts.project.deploy();
