--- bin/uninstall.js
@@ -1,3 +1,0 @@
-#!/usr/bin/env node
-
-require('../src/uninstall')();
